+++ rest_framework_dataclasses/types.py
@@ -4,9 +4,9 @@ from typing import Dict
 
 try:
     # Python 3.8 and later
-    from typing import Literal, Protocol
+    from typing import Final, Literal, Protocol
 except ImportError:
-    from typing_extensions import Literal, Protocol
+    from typing_extensions import Final, Literal, Protocol
 
 
 # Note that this doesn't actually work yet (https://stackoverflow.com/a/55240861)