+++ rest_framework_dataclasses/typing_utils.py
@@ -18,7 +18,7 @@ development seems to have stalled. Maybe in the future?
 import collections
 import typing
 
-from .types import Literal
+from .types import Final, Literal
 
 # typing._BaseGenericAlias was split-out from GenericAlias in Python 3.9
 if hasattr(typing, '_BaseGenericAlias'):
@@ -182,6 +182,24 @@ def get_literal_choices(tp: type) -> typing.List[typing.Union[str, bytes, int, b
     return values
 
 
+def is_final_type(tp: type) -> bool:
+    """
+    Test if the given type is a final type.
+    """
+    return tp is Final or get_origin(tp) is Final
+
+
+def get_final_type(tp: type) -> type:
+    """
+    Get the type that is made final.
+    """
+    if not is_final_type(tp):
+        raise ValueError('get_final_type() called with non-final type.')
+
+    args = get_args(tp)
+    return args[0] if len(args) > 0 else typing.Any
+
+
 def is_type_variable(tp: type) -> bool:
     """
     Test if the given type is a variable type.
@@ -189,15 +207,12 @@ def is_type_variable(tp: type) -> bool:
     return isinstance(tp, typing.TypeVar)
 
 
-def get_variable_types(tp: typing.TypeVar) -> typing.Optional[typing.List[type]]:
-    """
-    Get the possible types for a variable type.
-    """
-    return list(tp.__constraints__) if len(tp.__constraints__) > 0 else None
-
-
 def get_variable_type_substitute(tp: typing.TypeVar) -> type:
     """
     Get the substitute for a variable type.
     """
-    return typing.Union[tp.__constraints__] if len(tp.__constraints__) > 0 else typing.Any
+    if len(tp.__constraints__) > 0:
+        return typing.Union[tp.__constraints__]
+    if tp.__bound__ is not None:
+        return tp.__bound__
+    return typing.Any