+++ tests/test_functional.py
@@ -2,15 +2,22 @@ import copy
 import dataclasses
 import datetime
 import decimal
+import enum
 import typing
 import uuid
 
-from unittest import TestCase
+from unittest import TestCase, expectedFailure
 
-from rest_framework import fields, serializers
+from rest_framework import fields
 
 from rest_framework_dataclasses.serializers import DataclassSerializer
 from rest_framework_dataclasses.types import Literal
+
+
+class Gender(enum.Enum):
+    MALE = 'male'
+    FEMALE = 'female'
+    OTHER = 'other'
 
 
 @dataclasses.dataclass
@@ -26,6 +33,8 @@ class Person:
     name: str
     email: str
     phone: typing.List[str]
+    gender: typing.Optional[Gender] = None
+    length: typing.Optional[decimal.Decimal] = None
     pets: typing.Optional[typing.List[Pet]] = None
     birth_date: typing.Optional[datetime.date] = None
     favorite_pet: typing.Optional[Pet] = None
@@ -51,7 +60,8 @@ class PersonSerializer(DataclassSerializer):
 
     class Meta:
         dataclass = Person
-        fields = ('id', 'full_name', 'email', 'phone', 'pets', 'birth_date', 'favorite_pet', 'movie_ratings', 'slug', 'age')
+        fields = ('id', 'full_name', 'email', 'phone', 'gender', 'length', 'pets', 'birth_date', 'favorite_pet',
+                  'movie_ratings', 'slug', 'age')
         extra_kwargs = {
             'id': {'format': 'hex'},
             'phone': {'child_kwargs': {'max_length': 15}},
@@ -64,6 +74,12 @@ class FunctionalTestMixin:
     def test_serialize(self):
         serializer = self.serializer(instance=self.instance)
         self.assertDictEqual(serializer.data, {**self.representation, **self.representation_readonly})
+
+    def test_validated_data(self):
+        serializer = self.serializer(data=self.representation)
+        serializer.is_valid(raise_exception=True)
+
+        self.assertEqual(serializer.validated_data, self.instance)
 
     def test_create(self: TestCase):
         serializer = self.serializer(data=self.representation)
@@ -99,7 +115,9 @@ class PersonTest(TestCase, FunctionalTestMixin):
         id=uuid.UUID('28ee3ae5-480b-46bd-9ae4-c61cf8341b95'),
         name='Alice',
         email='alice@example.com',
+        length=decimal.Decimal('1.68'),
         phone=['+31-6-1234-5678', '+31-20-123-4567'],
+        gender=Gender.FEMALE,
         pets=[Pet(animal='cat', name='Milo', weight=decimal.Decimal('10.8')),
               Pet(animal='dog', name='Max', weight=decimal.Decimal('123.4'))],
         birth_date=datetime.date(1980, 4, 1),
@@ -110,7 +128,9 @@ class PersonTest(TestCase, FunctionalTestMixin):
         'id': '28ee3ae5480b46bd9ae4c61cf8341b95',
         'full_name': 'Alice',
         'email': 'alice@example.com',
+        'length': '1.68',
         'phone': ['+31-6-1234-5678', '+31-20-123-4567'],
+        'gender': 'female',
         'pets': [
             {'animal': 'cat', 'name': 'Milo', 'weight': '10.8'},
             {'animal': 'dog', 'name': 'Max', 'weight': '123.4'}
@@ -138,12 +158,46 @@ class EmptyPersonTest(TestCase, FunctionalTestMixin):
         'full_name': 'Alice',
         'email': 'alice@example.com',
         'phone': [],
+        'favorite_pet': None,
+    }
+    representation_readonly = {
+        'slug': 'Alice',
+        'length': None,
+        'gender': None,
         'pets': None,
         'birth_date': None,
         'age': None,
-        'favorite_pet': None,
         'movie_ratings': None
     }
-    representation_readonly = {
-        'slug': 'Alice'
-    }
+
+
+class PartialPersonTest(TestCase):
+    def test_update(self):
+        input_instance = Person(
+            id=uuid.UUID('28ee3ae5-480b-46bd-9ae4-c61cf8341b95'),
+            name='Alice',
+            email='alice@example.com',
+            phone=['+31-6-1234-5678', '+31-20-123-4567'],
+            favorite_pet=Pet('cat', 'Luna'),
+            pets=[Pet('dog', 'Bella'), Pet('cat', 'Luna')],
+        )
+
+        partial_representation = {
+            'full_name': 'Bob',
+            'email': 'bob@example.com',
+            'favorite_pet': {'name': 'Molly'},
+            'pets': [{'animal': 'cat', 'name': 'Molly'}],
+        }
+
+        expected_output = dataclasses.replace(input_instance,
+                                              name='Bob',
+                                              email='bob@example.com',
+                                              favorite_pet=Pet(animal='cat', name='Molly'),
+                                              pets=[Pet('cat', 'Molly')])
+
+        serializer = PersonSerializer(instance=input_instance, data=partial_representation, partial=True)
+        serializer.is_valid(raise_exception=True)
+        output_instance = serializer.save()
+
+        self.assertIs(output_instance, input_instance)
+        self.assertEqual(output_instance, expected_output)